When adding metadata, sometimes two instances of it would show up after adding, but (#1201)
only one was on the server (refreshing showed one). Combining the useEffect statements
seems to have fixed that.
At some point we should remove the complex if logic included with the jsx. We might want
to have separate components for dataset and files and a common one for the actual entries.